refactor: remove filthy spaghetti when starting CACE phase from FA/RA
Instead of using function calls with Enter_Method, RINASim for the most
part uses so-called listeners, which are subscribed to various signals
in various scopes. This adds a large amount of technical overhead for
anyone wanting to understand any part of this codebase. With this
commit, FA now does not emit "finished creating management flow"
signals, but instead calls the startCACE Enrollment method directly.

Another reason that these changes should be made all through the code
base, is the fact that while singals can be used for event polling, they
should more frequently be used for recording statistics.

A big design flaw in RINASim is the fact that all instances of IPC
Processes have their own set of listeners that all subscribe to signals
that are propagated to the host. For potential large scale simulation
scenarios, this creates an extreme processing overhead. If RINASim is
ever going to be seen as a serious implementation of RINA, this is on
the forefront of the issues that need to be fixed.