refactor: major flow allocation code improvements
Signals in both FA and FAI have gotten a redux, as they now a naming
scheme that follow some convention. Also, FAListeners is completely
removed in favour of allowing FA to listen to signals directly. This
will likely change more as I rework the API.

Also created a prototype function for reworking the allocateRequest
primitive. To allow for easier development and a greater degree of
extensibility, this should be the top priority, not only for EthShimDIF,
but also for potential future shim modules.