test: update fingerprints to more(?) correct values
Need to update to use for automated tests. Most of the new fingerprints
are in line with the fingerprints of the latest RINASim commit in the
main repo (kvetak/RINASim), but some examples appears to have been fixed
by the changes I have introduced (among them, all SimpleLS tests).

Still a lot of broken examples that haven't been iteratively updated.