feat(FA): check for connection to remote IPCP on forward oc M_CREATE
Forwarded creation requests just assumed that the IPCP was already
enrolled and connected. This had consequences for the ethernet shim DIF,
since there was a function that specifically waited for creation
of an N-1 non-management flow (using `flow->isManagementFlow`). The shim
IPCP only allows one flow between each set of end-points, and so the
first flow that allocates will always be management, which made the
forwarding incompatible. This commit fixes the issue by first checking
if a management flow is required to reach given IPCP.

Also set `pendingFlows` to be a queue instead of a list, since the list
was operated on exactly as a queue would be. Probably means nothing for
performance, as cache locality was thrown out the window with the use of
pointers for everything.