fix: avoid trying to change ownership of object in event heap
CACE module previously tried to take ownership of a passed object since
this is required before sending. With this, the message is duplicated in
CACE instead, and the objects are constructed on the stack to avoid
memory leaks. This is by no means a good fix, but still magnitudes
better than just taking ownership in CACE. CACE should either be
directly responsible for constructing relevant messages (or else it
serves no other purpose than visualisation), or just be removed.

Should also fix some examples that previously did not work.