feat(shim): add MessageDispatcher and InterfaceTable to IPCProcessShim
When several Ethernet Interfaces exist on a host have the same name, a
runtime exception is thrown, regardless of which submodule they are in.
This is because InterfaceTables checks if an entry with the name already
exists in the table, instead of checking entire path. Since there is no
good way of giving the interfaces individual names within each shim IPC
Process, the interface table module is now moved into the shim IPC
Process. It is required to access the MAC address.

MessageDispatcher is used in the new Packet API of INET, and has you
explicitly add new protocols. Thiss will allow a RINA ethtype to be
added explicitly as well, which was not done before as part of the
ProtocolGroup stuff in INET 4.